# openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "CaseAuth API",
    "version": "v1"
  },
  "paths": {
...
    "/api/cases/{caseId}/decisions": {
      "get": {
        "tags": [
          "Decisions"
        ],
        "operationId": "Decisions_List",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionResponse"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionResponse"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DecisionResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Decisions"
        ],
        "operationId": "Decisions_Create",
        "parameters": [
          {
            "name": "caseId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "If-Match",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDecisionRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDecisionRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DecisionResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDecisionRequest": {
        "type": "object",
        "properties": {
          "outcome": {
            "$ref": "#/components/schemas/DecisionOutcome"
          },
          "aiReviewId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DecisionOutcome": {
        "enum": [
          "Approved",
          "Rejected",
          "Escalated"
        ],
        "type": "string"
      },
      "DecisionResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "caseId": {
            "type": "string",
            "format": "uuid"
          },
          "outcome": {
            "$ref": "#/components/schemas/DecisionOutcome"
          },
          "aiReviewId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "decidedByUserId": {
            "type": "string",
            "nullable": true
          },
          "decidedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "DevUser": {
        "type": "apiKey",
        "description": "Development-only auth: set to a seeded username (analyst1, analyst2, supervisor).",
        "name": "X-Dev-User",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "DevUser": [ ]
    }
  ]
}